Add local encryption key validation
When user creates an encryption key, we encrypt a challenge string with their encryption key to their metadata. When user tries validating the key, we try decoding the encrypted challenge with their key. This way the actual key is never sent to the server. This step is needed to save the key to IndexedDB, since we don't store the key in a database, it needs to be manually set every time by the user.

Changes:
- Added validation screen to authentication flow
- Added "onKeyDown" prop to TextInput component
- Changed the page structure, the main file manager is now located at "/home" and landing page is "/"
- Changed "encryptionKey" field in user metadata to contain the encrypted challenge with the needed props for decrypting
- Different authentication states now show correctly
- Encrypting, Decrypting and IndexedDB now return values to indicate if the operation was successful
- Added temporary landing page