Fix error on account deletion
When user deletes their account, all their files are also deleted. This triggers the "modifyfile" function, but because the path to the user's modified file doesn't exist anymore, it used to throw an error. This commit fixes the error by checking if the user still exists before trying to access the file.